Merge with main

+++ code/game/machinery/OpTable.dm
@@ -3,15 +3,22 @@
 	desc = "Used for advanced medical procedures."
 	icon = 'icons/obj/surgery.dmi'
 	icon_state = "optable-idle"
+
+	layer = TABLE_LAYER
 	density = 1
 	anchored = 1.0
 	use_power = 1
 	idle_power_usage = 1
 	active_power_usage = 5
-	var/mob/living/carbon/human/victim = null
+
+	var/mob/living/carbon/victim = null
 
 	var/obj/machinery/computer/operating/computer = null
+	can_buckle = TRUE
+	buckle_dir = SOUTH
+	buckle_lying = TRUE //bed-like behavior, forces mob.lying = buckle_lying if != -1
 
+	var/y_offset = 0
 /obj/machinery/optable/New()
 	..()
 	for(var/dir in list(NORTH,EAST,SOUTH,WEST))
@@ -20,7 +27,7 @@
 			computer.table = src
 			break
 //	spawn(100) //Wont the MC just call this process() before and at the 10 second mark anyway?
-//		process()
+//		Process()
 
 /obj/machinery/optable/ex_act(severity)
 
@@ -36,16 +43,24 @@
 				return
 		if(3.0)
 			if (prob(25))
-				src.density = 0
+				density = 0
 		else
 	return
 
 /obj/machinery/optable/attack_hand(mob/user as mob)
+	if (user.incapacitated(INCAPACITATION_DEFAULT))
+		return
+	if (victim)
+		user_unbuckle_mob(user)
+		return
 	if (HULK in usr.mutations)
 		visible_message(SPAN_DANGER("\The [usr] destroys \the [src]!"))
-		src.density = 0
+		density = 0
 		qdel(src)
-	return
+
+/obj/machinery/optable/unbuckle_mob()
+	. = ..()
+	check_victim()
 
 /obj/machinery/optable/CanPass(atom/movable/mover, turf/target, height=0, air_group=0)
 	if(air_group || (height==0)) return 1
@@ -56,17 +71,18 @@
 		return 0
 
 /obj/machinery/optable/proc/check_victim()
-	if(locate(/mob/living/carbon/human, src.loc))
-		var/mob/living/carbon/human/M = locate(/mob/living/carbon/human, src.loc)
-		if(M.lying)
-			src.victim = M
+	if (istype(buckled_mob,/mob/living/carbon))
+		victim = buckled_mob
+		if(ishuman(buckled_mob))
+			var/mob/living/carbon/human/M = buckled_mob
 			icon_state = M.pulse() ? "optable-active" : "optable-idle"
-			return 1
-	src.victim = null
+		return 1
+
+	victim = null
 	icon_state = "optable-idle"
 	return 0
 
-/obj/machinery/optable/process()
+/obj/machinery/optable/Process()
 	check_victim()
 
 /obj/machinery/optable/proc/take_victim(mob/living/carbon/C, mob/living/carbon/user as mob)
@@ -74,20 +90,17 @@
 		user.visible_message("[user] climbs on \the [src].","You climb on \the [src].")
 	else
 		visible_message(SPAN_NOTICE("\The [C] has been laid on \the [src] by [user]."), 3)
+		if (user.pulling == C)
+			user.stop_pulling() //Lets not drag your patient off the table after you just put them there
 	if (C.client)
 		C.client.perspective = EYE_PERSPECTIVE
 		C.client.eye = src
-	C.resting = 1
-	C.loc = src.loc
+	C.loc = loc
 	for(var/obj/O in src)
-		O.loc = src.loc
-	src.add_fingerprint(user)
-	if(ishuman(C))
-		var/mob/living/carbon/human/H = C
-		src.victim = H
-		icon_state = H.pulse() ? "optable-active" : "optable-idle"
-	else
-		icon_state = "optable-idle"
+		O.loc = loc
+	add_fingerprint(user)
+	buckle_mob(C)
+
 
 /obj/machinery/optable/MouseDrop_T(mob/target, mob/user)
 
@@ -116,9 +129,17 @@
 /obj/machinery/optable/proc/check_table(mob/living/carbon/patient as mob)
 	check_victim()
 	if(victim)
-		usr << SPAN_WARNING("\The [src] is already occupied!")
+		to_chat(usr, SPAN_WARNING("\The [src] is already occupied!"))
 		return 0
 	if(patient.buckled)
-		usr << SPAN_NOTICE("Unbuckle \the [patient] first!")
+		to_chat(usr, SPAN_NOTICE("Unbuckle \the [patient] first!"))
 		return 0
 	return 1
+
+/obj/machinery/optable/post_buckle_mob(mob/living/M as mob)
+	if(M == buckled_mob)
+		M.pixel_y = y_offset
+	else
+		M.pixel_y = 0
+
+	check_victim()
